\name{AppliedPredictiveModeling-package}
\alias{AppliedPredictiveModeling-package}
\alias{AppliedPredictiveModeling}
\docType{package}
\title{
What the package does (short line)
 package title 
}
\description{
More about what it does (maybe more than one line)
 A concise (1-5 lines) description of the package
}
\details{
\tabular{ll}{
Package: \tab AppliedPredictiveModeling\cr
Type: \tab Package\cr
Version: \tab 1.1-001\cr
Date: \tab 2013-05-29\cr
License: \tab GPL\cr
}
 An overview of how to use the package, including the most important functions 
}
\author{
Max Kuhn

Maintainer: Max Kuhn <mkuhn@gmail.com>
}
\references{
 Kuhn M and Johnson K (2013) Applied Predictive Modeling, Springer, NY
}
\keyword{ package }
