% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackages.R
\name{getPackages}
\alias{getPackages}
\title{Install Packages for Each Chapter}
\usage{
getPackages(chapter, ...)
}
\arguments{
\item{chapter}{an integer vector (or character versions of the integer) for
the chapter number. See Details below:}

\item{...}{options to pass to \code{\link[utils]{install.packages}}}
}
\description{
This function identifies the physical location on the user's computer where
the chapter R scripts are located.
}
\details{
Chapter names and packages. about dependencies.
}
\examples{

\dontrun{
getPackages(2)
getPackages(2:3)
getPackages("4")
}

}
\author{
Max Kuhn
}
\keyword{utilities}
