% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppliedPredictiveModeling-package.R
\docType{data}
\name{logisticCreditPredictions}
\alias{logisticCreditPredictions}
\title{Logistic Regression Predictions for the Credit Data}
\value{
A data frame with columns \item{Bad}{The predicted class probability
for bad credit. } \item{Good}{The predicted class probability for good
credit.} \item{pred}{The predicted class. } \item{obs}{The observed class }
}
\description{
add some notes
}
\examples{

## show code to make the predictions

}
\keyword{datasets}
