% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptLocation.R
\name{scriptLocation}
\alias{scriptLocation}
\title{Find Chapter Script Files}
\usage{
scriptLocation()
}
\description{
This function identifies the physical location on the user's computer where
the chapter R scripts are located.
}
\examples{

scriptLocation()

}
\author{
Max Kuhn
}
\keyword{utilities}
