% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twoClassData}
\alias{twoClassData}
\alias{classes}
\title{Two Class Example Data}
\value{
\item{predictors}{data frame of two predictors} \item{classes}{a
factor vector of class labeled}
}
\description{
These data contain two predictors measured for 208 samples. Of these, 111
samples are labeled as \code{Class1} and the remaining 97 are \code{Class2}.
}
\examples{

data(twoClassData)

library(lattice)
xyplot(PredictorB ~ PredictorA, 
       data = predictors, 
       groups = classes, 
       auto.key = TRUE)


}
\keyword{datasets}
