\name{abalone}
\docType{data}
\alias{abalone}
\title{Abalone Data}
\description{
The Abalone data consist of data from 4177 abalones. The data consist of measurements of the type (male, female and infant), the longest shell measurement, the diameter, height and several weights (whole, shucked, viscera and shell). The outcome is the number of rings. The age of the abalone is the number of rings plus 1.5. 

The data are taken from the UCI database (\url{http://archive.ics.uci.edu/ml/datasets/Abalone}).
}

\usage{data(abalone)}

\value{
  \item{abalone}{a data frame with 4177 rows and 9 columns}
  }

\examples{
data(abalone)
}

\keyword{datasets}  

