% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppliedPredictiveModeling-package.R
\name{apm-internal}
\alias{apm-internal}
\alias{lowerp}
\alias{upperp}
\title{Internal Functions}
\arguments{
\item{\dots}{optional arguments to pass to internal functions}
}
\description{
Internal functions
}
\author{
Max Kuhn
}
\keyword{internal}
