\name{AppliedPredictiveModeling-package}
\alias{AppliedPredictiveModeling-package}
\alias{AppliedPredictiveModeling}
\docType{package}
\title{
Data, Functions and Scripts for 'scriptLocation'
}
\description{
This package can be used to reproduce the analyses in the text. Scripts for each chapter are located in the "chapters" directory. Use \code{scriptLocation()} to find their exact location.
}
\details{
\tabular{ll}{
Package: \tab AppliedPredictiveModeling\cr
Type: \tab Package\cr
Version: \tab 1.1-1\cr
Date: \tab 2013-05-29\cr
License: \tab GPL\cr
}
}
\author{
Max Kuhn

Maintainer: Max Kuhn <mkuhn@gmail.com>
}
\references{
 Kuhn M and Johnson K (2013) Applied Predictive Modeling, Springer, NY
}
\keyword{ package }
