\name{AppliedPredictiveModeling-package}
\alias{AppliedPredictiveModeling-package}
\alias{AppliedPredictiveModeling}
\docType{package}
\title{
What the package does (short line)
 package title 
}
\description{
More about what it does (maybe more than one line)
 A concise (1-5 lines) description of the package
}
\details{
\tabular{ll}{
Package: \tab AppliedPredictiveModeling\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-12-20\cr
License: \tab GPL-2\cr
}
 An overview of how to use the package, including the most important functions 
}
\author{
Max Kuhn

Maintainer: Max Kuhn <mkuhn@gmail.com>
}
\references{
 Literature or other references for background information 
}
\keyword{ package }
\seealso{
 Optional links to other man pages, e.g. 
 
}
\examples{
## simple examples of the most important functions 
}
