\name{logisticCreditPredictions}
\docType{data}
\alias{logisticCreditPredictions}
\title{Logistic Regression Predictions for the Credit Data}
\description{
add some notes
}

\usage{data(solubility)}

\value{
A data frame with columns
  \item{Bad}{The predicted class probability for bad credit. }
  \item{Good}{The predicted class probability for good credit.}
  \item{pred}{The predicted class. }
  \item{obs}{The observed class }
  }



\examples{
## show code to make the predictions
}

\keyword{datasets}  

