\name{scriptLocation}
\alias{scriptLocation}
\title{
Find Chapter Script Files
}
\description{
This function identifies the physical location on the user's computer where the chapter R scripts are located.
}
\usage{
scriptLocation()
}

\author{
Max Kuhn
}

\examples{
scriptLocation()
}

\keyword{utilities}
